[FIX]: 피드백 반영, PriorityFromHeader 중복 제거 — pkg/queue 로 단일 출처 이관
gemini #3278202670 (Medium) DRY 위반 반영. Parser / Validate / Enrich 의 동일 함수
3개를 pkg/queue.PriorityFromHeader 로 통합.

변경:
- pkg/queue/priority_header.go (신규):
  - PriorityFromHeader(headers) int — 단일 출처
  - PriorityHeaderKey 상수 ("priority")
- internal/processor/{parser,validate,enrich}/worker/:
  - 각 패키지의 PriorityFromHeader 정의 제거 + 이관 안내 주석
  - 호출처를 queue.PriorityFromHeader 로 변경
  - strconv import 제거 (parser/zset_intake.go)
- test/pkg/queue/priority_header_test.go (신규): 9 cases 단일 테스트
- 각 worker package 의 중복 TestPriorityFromHeader_Mapping 제거

gemini #3278202667 (HIGH) at-least-once 보장은 이미 c10b4b5 / Enrich PR 본문에
구현됨 (enqueueRetry 실패 → DLQ fallback → 실패 시 commit X). timeout_ms 헤더
계승도 BuildRetryJob 에 이미 포함. context.WithoutCancel 패치는 Parser/Validate
와의 일관성을 고려해 본 PR scope 외로 분리.
